Release 2.1.15 (#186)
* Update a lot of debugging out put to DEBUG_INFO.  It was inconsistent.

* Add Modbus (#156).   v2.1.10

Version 2.1.10, merging in Modbus functionality. This includes:

- Initial Modbus support. Very initial!
- complete refactoring of the byte order handling.
- fixes for a memory found in socket_connect_tcp() when a host name is used. Did not leak in Windows.
- updates and warnings for deprecated tag attributes.  Just a start.
- support for getting/setting debug level using 'debug' attribute rather than 'debug_level'.

* Fix for #155, handle negative timeout.

* updating tag count.

* Fix #157, Micro800 regression (#158)

* updating tag count.

* Fix #157, Micro800 connection regression.

* Add Micro800 tests.

* Normalize error returns for use in automated testing.

* Updating tests.

* Added termination flag and signal handling.

* Fix micro800 regression (#159)

* updating tag count.

* Fix #157, Micro800 connection regression.

* Add Micro800 tests.

* Normalize error returns for use in automated testing.

* Updating tests.

* Added termination flag and signal handling.

* Run all tests on both x64 and x86

* Fix #160.  Return correct values on error.

* Remove unused tag_set_busy_flag() function.

* Add modbus (#161)

Updating Modbus code to add inactivity timeout and automatic reconnection.

* Add support for Long Int for PCCC. (#164)

* Add port handling to AB PLCs. Fix #165. (#166)

* Add Omron NJ/NX, PR #153 (#168)

OMRON NJ/NX Series Ethernet/IP initial support.


* Add branch name in artifact name.

* Removing attempt at putting branch into artifact name.

* platform: Fix resource and memory leaks in socket_connect_tcp (#169)

Fix socket and memory leaks in error path in socket_connect_tcp(), thanks to @jubnzv!

* Limit elem_count to 1 on OMRON PLCs. (#171)

* Annotate fall through to stop compiler warning.

* Add Omron simulator/tests.

* Add Modbus (#156).   v2.1.10

Version 2.1.10, merging in Modbus functionality. This includes:

- Initial Modbus support. Very initial!
- complete refactoring of the byte order handling.
- fixes for a memory found in socket_connect_tcp() when a host name is used. Did not leak in Windows.
- updates and warnings for deprecated tag attributes.  Just a start.
- support for getting/setting debug level using 'debug' attribute rather than 'debug_level'.

* Fix micro800 regression (#159)

* updating tag count.

* Fix #157, Micro800 connection regression.

* Add Micro800 tests.

* Normalize error returns for use in automated testing.

* Updating tests.

* Added termination flag and signal handling.

* Run all tests on both x64 and x86

* Fix #160.  Return correct values on error.

* Add modbus (#161)

Updating Modbus code to add inactivity timeout and automatic reconnection.

* Add Omron NJ/NX, PR #153 (#168)

OMRON NJ/NX Series Ethernet/IP initial support.


* Add branch name in artifact name.

* Removing attempt at putting branch into artifact name.

* Limit elem_count to 1 on OMRON PLCs. (#171)

* Annotate fall through to stop compiler warning.

* Add Omron simulator/tests.

* Port ab_server to Windows (#173)

Redo ab_server for Windows.

* Merge auto read/write functionality from auto_read_write branch. (#174)

* Changing tag status to int8_t and reorganized tag elements to reduce struct size.

* Adding test program for auto sync.

* Adding test program for auto sync.

* Update version number.

* Set the read_in_flight flag when kicking off initial read.

* Add first cut of automatic read/write.

* Fix commented out mutex_lock/unlock.

* Missed a commit of code to clear the in flight flags.

* Fixed debug level.

* Removed debug level.

* More clean up around read/write_in_flight flags.

* Changing tag status to int8_t and reorganized tag elements to reduce struct size.

* Adding test program for auto sync.

* Adding test program for auto sync.

* Set the read_in_flight flag when kicking off initial read.

* Add first cut of automatic read/write.

* Fix commented out mutex_lock/unlock.

* Missed a commit of code to clear the in flight flags.

* Fixed debug level.

* Removed debug level.

* More clean up around read/write_in_flight flags.

* Changing tag status to int8_t and reorganized tag elements to reduce struct size.

* Adding test program for auto sync.

* Adding test program for auto sync.

* Set the read_in_flight flag when kicking off initial read.

* Add first cut of automatic read/write.

* Fix commented out mutex_lock/unlock.

* Missed a commit of code to clear the in flight flags.

* Fixed debug level.

* Removed debug level.

* More clean up around read/write_in_flight flags.

* Add first cut of automatic read/write.

* More clean up around read/write_in_flight flags.

* Added indentation for clarity.

* Fixed up timing to make automatic reads try hard to be exactly on schedule.

* Fixed typo in tag attribute string.

* Update version to 2.1.12 for auto tags.

* Need to rewrite test_auto_sync for Windows.

* Add basic PLC/5 support to ab_server.

* Added SLC 500/Micrologix support to ab_server.

* Add PLC/5 and SLC 500 tests to CI.

* Add PCCC PLC emulation support to ab_server. (#175)

* Add basic PLC/5 support to ab_server.

* Added SLC 500/Micrologix support to ab_server.

* Add PLC/5 and SLC 500 tests to CI.

* Updated README to note support for .Net Core, .Net Framework.

* Fix #176.   Make connection sequence ID non-repeating. (#177)

* Fix large tag breakage due to Omron merge. (#178)

* Fix breakage from Omron merge.  Long tags did not work.

* Bump version for Omron/large tag fix.

* Support multiple ForwardOpen attempts.   Add FO test and large tag tests to CI.  Add CI tests to macOS. (#180).  Fix #176.

* Add FO duplicate connection ID error handling.
* Do multiple ForwardOpen attempts.  Update CI to test FO.
* Add large tag tests to CI.
* Add CI tests to macOS.

* CI tests for Windows. (#181)

Add CI tests for Windows.

* Add fix for #183. (#184)

* Fix up tag path for real PLC address.

* Add guard to prevent memory allocation of zero bytes.

* Add additional guards for PLC host address being zero length or missing.

* Add guards and additional logic to most string functions to treat zero-length and NULL strings the same.

* Bump version for release.

* Fix #182, timestamp in Windows logging was wrong (#185)

* Tweaked time_ms() and fixed off-by-one for the month on date output in logs.

+++ src/protocols/ab/eip_plc5_dhp.c
@@ -186,7 +186,7 @@ int tag_tickler(ab_tag_p tag)
     if(tag->read_in_progress) {
         pdebug(DEBUG_SPEW, "Read in progress.");
         rc = check_read_status(tag);
-        tag->status = rc;
+        tag->status = (int8_t)rc;
 
         /* check to see if the read finished. */
         if(!tag->read_in_progress) {
@@ -199,7 +199,7 @@ int tag_tickler(ab_tag_p tag)
     if(tag->write_in_progress) {
         pdebug(DEBUG_SPEW, "Write in progress.");
         rc = check_write_status(tag);
-        tag->status = rc;
+        tag->status = (int8_t)rc;
 
         /* check to see if the write finished. */
         if(!tag->write_in_progress) {
